hopefully fix things

+++ Brickwell/Brickwell/Models/Order.cs
@@ -14,7 +14,7 @@ namespace Brickwell.Models
         public required string day_of_week { get; set; }
         public required int time { get; set; }
         public required string entry_mode { get; set; }
-        public required int amount { get; set; }
+        public required decimal amount { get; set; }
         public required string type_of_transaction { get; set; }
         public required string country_of_transaction { get; set; }
         public required string shipping_address { get; set; }
